+++ R/Set_LogicalSet.R
@@ -1,7 +1,6 @@
 #' @name LogicalSet
 #' @title Set of Logicals
 #' @description The `LogicalSet` is defined as the [Set] containing the elements `TRUE` and `FALSE`.
-#' @family sets
 #'
 #' @examples
 #' l <- LogicalSet$new()
@@ -15,8 +14,10 @@ LogicalSet <- R6::R6Class("LogicalSet",
     #' @details The Logical set is the set containing `TRUE` and `FALSE`.
     #' @return A new `LogicalSet` object.
     initialize = function() {
+      warning("Deprecated. In the future please use Logicals$new(). This will be removed in v0.4.0.")
       private$.elements <- list(TRUE, FALSE)
       private$.str_elements <- c("TRUE", "FALSE")
+      private$.multiplicity <- list("TRUE" = 1, "FALSE" = 1)
       private$.class <- "logical"
       private$.properties <- Properties$new(closure = "closed", cardinality = 2)
       private$.lower <- TRUE