+++ NEWS.md
@@ -1,3 +1,15 @@
+# set6 0.2.0
+
+* UniversalSet renamed Universal, old class will be removed in v0.4.0.
+* LogicalSet renamed Logicals, old class will be removed in v0.4.0.
+* `Complex` now inherits from `Set`, incorrect methods for `isSubset, equals` have been removed.
+* Add `Multiset` for sets with non-unique elements but no ordering
+* Small speed improvements in `Tuple` and `FuzzyTuple`
+* For consistency most methods now return a `list` unless single elements requested
+* Printing of `ConditionalSet` is fixed
+* `Rationals` and child-classes now error on calls to `contains, isSubset, equals` as any prior results were likely wrong/misleading
+* Removed erroneous complex boundaries in `Interval` class
+
 # set6 0.1.8
 
 * Patch for R-devel